use a new controller and completely exempt it if property doesnt exist
