+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -2,6 +2,7 @@ uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.FilterHelper
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.DataApiAsyncController
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.DataApiSyncController
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ReportDefinitionController
+uk.gov.justice.digital.hmpps.digitalprisonreportinglib.missingReport.MissingReportSubmissionController
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.DashboardDefinitionController
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.config.DprOpenApiConfiguration
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.config.DefinitionGsonConfig
