+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/integration/MissingReportIntegrationTest.kt
@@ -26,7 +26,7 @@ class MissingReportNoDatasourceIntegrationTest : IntegrationTestBase() {
       .bodyValue("a reason")
       .headers(setAuthorisation(roles = listOf(authorisedRole), user = "foo"))
       .exchange()
-      .expectStatus().isEqualTo(501)
+      .expectStatus().isEqualTo(404)
   }
 }
